Move docker_lookup_request to a separate file
