+++ userspace/libsinsp/container_engine/docker.h
@@ -37,6 +37,7 @@ limitations under the License.
 #include "container.h"
 #include "container_info.h"
 
+#include "container_engine/docker/lookup_request.h"
 #include "container_engine/container_engine_base.h"
 #include "container_engine/sinsp_container_type.h"
 #include "container_engine/wmi_handle_source.h"
@@ -46,38 +47,6 @@ class sinsp_threadinfo;
 
 namespace libsinsp {
 namespace container_engine {
-
-struct docker_lookup_request
-{
-	docker_lookup_request() :
-		request_rw_size(false)
-	{}
-
-	docker_lookup_request(const std::string& container_id_value,
-			      bool rw_size_value) :
-		container_id(container_id_value),
-		request_rw_size(rw_size_value)
-	{}
-
-	bool operator<(const docker_lookup_request& rhs) const
-	{
-		if(container_id != rhs.container_id)
-		{
-			return container_id < rhs.container_id;
-		}
-
-		return request_rw_size < rhs.request_rw_size;
-	}
-
-	bool operator==(const docker_lookup_request& rhs) const
-	{
-		return container_id == rhs.container_id &&
-		       request_rw_size == rhs.request_rw_size;
-	}
-
-	std::string container_id;
-	bool request_rw_size;
-};
 
 class docker_async_source : public sysdig::async_key_value_source<docker_lookup_request, sinsp_container_info>
 {
